app/requirements.txt
Flask==2.0.1
tweepy==4.5.0
Werkzeug==2.2.2
python-dotenv==1.0.1